Divide classes in separate files
